refactor [1/2]
* remove blackhand
  was not useful at all
* remove the persisted dictionary layer (was used for dictionary compressing keys)
  seems it was just adding complexity
* move forward index into per "segment" (day) directory
  this allows us to use 32 bits instead of 64 for the inverted index
* remove storing very different form of the data spec.PersistedMetadata
  simply store spec.Metadata, but use values of KV instead of poitners &KV
* remove storing the kafka offset
  was not useful at all
* remove storing the payload
  was not useful at all
* make all queries always decode the metadata
* make time fixed64 instead of varint
* simplify the forward storage format to be only [len, checksum, data]
  having foreign_id, foreign_type, event_type, len, checksumHeader, checksumData, data
  was just complicated for no good reasion, since we almost always need to decode
  the data, and adding more stuff to the header just takes space and time

+++ jubei/disk/dictionary_test.go
@@ -14,7 +14,7 @@ func TestDictionary(t *testing.T) {
 	}
 	defer os.RemoveAll(dir)
 
-	cases := map[string]uint64{}
+	cases := map[string]uint32{}
 	for k := 0; k < 10; k++ {
 		d, err := NewPersistedDictionary(dir)
 		if err != nil {
@@ -63,5 +63,4 @@ func TestDictionary(t *testing.T) {
 		}
 		d.Close()
 	}
-
 }
